refactor: all placeholder commands in commands/ directory
