+++ commands/configBit.js
