+++ commands/initBit.js
