+++ commands/tokenBit.js
