--- initBit.js
@@ -1,5 +1,0 @@
-function initBit() {
-  console.log("Initialize bit");
-}
-
-module.exports = { initBit };
